fix: 4단계 피드백 반영
