+++ src/main/java/racing/Simulator.java
@@ -1,14 +1,13 @@
 package racing;
 
+import static racing.model.Cars.getWinners;
+
+import java.util.ArrayList;
+import java.util.List;
+import racing.model.Cars;
 import racing.model.RacingCar;
 import racing.view.InputView;
 import racing.view.ResultView;
-
-import java.util.ArrayList;
-import java.util.Arrays;
-import java.util.List;
-import java.util.stream.Collectors;
-import java.util.stream.Stream;
 
 public class Simulator {
     private static final int BEGINNING = 0;
@@ -16,43 +15,23 @@ public class Simulator {
     static String[] carName;
     static int carCount;
     static int raceRound;
-    static List<RacingCar> cars = new ArrayList<>();
 
     public static void main(String[] args) throws IllegalAccessException {
         carName = InputView.getCarName();
         carCount = carName.length;
         raceRound = InputView.getRaceRound();
+        List<RacingCar> racingCars = new ArrayList<>();
         for (int i = 0; i < carCount; i++) {
-            cars.add(new RacingCar(BEGINNING, carName[i]));
+            racingCars.add(new RacingCar(BEGINNING, carName[i]));
         }
+        Cars car = new Cars(racingCars);
         System.out.println();
         System.out.println("실행 결과");
         for (int i = 0; i < raceRound; i++) {
-            raceAllCars();
+            car.raceAllCars();
             System.out.println();
         }
-        ResultView.showWinners(getWinners(cars));
-    }
-
-    public static List<RacingCar> getWinners(List<RacingCar> cars) {
-        return cars.stream()
-                .filter(car -> car.isSameProgress(getTopSpeed(cars)))
-                .collect(Collectors.toList());
-    }
-
-    public static int getTopSpeed(List<RacingCar> cars) {
-        int max = 0;
-        for (RacingCar car : cars) {
-            max = car.max(max);
-        }
-        return max;
-    }
-
-    private static void raceAllCars() {
-        for (RacingCar car : cars) {
-            car.race(car.getRandomInt());
-            ResultView.showRaceProgress(car);
-        }
+        ResultView.showWinners(getWinners(racingCars));
     }
 
 }
