+++ src/main/java/racing/model/Calculator.java
@@ -1,4 +1,4 @@
-package calculator;
+package racing.model;
 
 import java.util.regex.Matcher;
 import java.util.regex.Pattern;
